Fix type casts

+++ fw/analog_turntable.c
@@ -21,9 +21,9 @@ int8_t analog_turntable_poll(analog_turntable* self, uint32_t current_value) {
 
   // is the current value sufficiently far away from the center?
   uint8_t direction = 0;
-  if (delta >= (int32_t)self->deadzone) {
+  if (delta >= (int8_t)self->deadzone) {
     direction = 1;
-  } else if (delta <= -(int32_t)self->deadzone) {
+  } else if (delta <= -(int8_t)self->deadzone) {
     direction = -1;
   }
